Major firebase code refactor. Moved !ask to utility.
