+++ commands/a_boothing/remove-private-boothing.js
@@ -1,6 +1,5 @@
 // Discord.js commando requirements
 const { Command } = require('discord.js-commando');
-const firebaseServices = require('../../firebase-services');
 const discordServices = require('../../discord-services');
 
 // Command export
