+++ commands/h_workshop/ask-for-help.js
@@ -1,6 +1,8 @@
 // Discord.js commando requirements
 const { Command } = require('discord.js-commando');
-const firebaseServices = require('../../firebase-services');
+const firebaseServices = require('../../firebase-services/firebase-services');
+
+const firebaseWorkshops = require('../../firebase-services/firebase-services-workshops');
 const discordServices = require('../../discord-services');
 
 // Command export
@@ -32,7 +34,7 @@ module.exports = class AskForHelp extends Command {
 
             workshop = workshop.slice(0, workshop.indexOf('text') - 1);
 
-            var status = await firebaseServices.addToTAHelp(workshop, username);
+            var status = await firebaseWorkshops.addHacker(workshop, username);
 
             // If the user is alredy in the waitlist then tell him that
             if (status === firebaseServices.status.HACKER_IN_USE) {
@@ -46,7 +48,7 @@ module.exports = class AskForHelp extends Command {
 
                 // get ta console
                 var channel = await message.guild.channels.cache.find(channel => channel.name === workshop + '-ta-console');
-                var number = await firebaseServices.leftInTAHelpList(workshop);
+                var number = await firebaseWorkshops.leftToHelp(workshop);
                 channel.send('There are: ' + number + ' hackers waiting in line!');
             }
         }
